refactor: move path/session-dir helpers into module
